Change overlay to support multiple overlays
- Change so overlays must be given a name, as for frames or dialogs
- Add parameter "zorder" to overlay(). The overlays will be drawn according to the zOrder (high zOrder overlays are drawn over low zOrder overlays)
- Move HTMLJFXPanel WebView code to HTMLWebViewManager
- Separate HTMLOverlay into HTMLOverlayManager and HTMLOverlayPanel
- Feature discussed in #1425
